Album duration is now an int instead of a long, as int is large enough.

+++ mediaprovider/core/src/main/java/com/simplecityapps/mediaprovider/model/Album.kt
@@ -8,7 +8,7 @@ class Album(
     val albumArtistId: Long,
     val albumArtistName: String,
     val songCount: Int,
-    val duration: Long,
+    val duration: Int,
     var year: Int
 ) : Serializable {
